+++ lib/scss_lint/reporter/xml_reporter.rb
@@ -8,7 +8,9 @@ module SCSSLint
         output << "<file name='#{filename}'>"
 
         file_lints.each do |lint|
-          output << "<issue line='#{lint.line}' severity='warning' reason='#{lint.description}' />"
+          output << "<issue line='#{lint.line}' " <<
+                           "severity='#{lint.severity}' " <<
+                           "reason='#{lint.description}' />"
         end
 
         output << '</file>'