+++ CHANGELOG.md
@@ -6,6 +6,11 @@
   started with a shorthandable property name could report false positives
 * Improve message reported by `Shorthand` linter to display the shortest
   possible form
+* Syntax errors are now differentiated from lint warnings in `scss-lint`'s
+  output by using `E` and `W` respectively.
+* `NoZeroBeforeDecimal` linter has been renamed to `LeadingZero`
+* Fix bug where `LeadingZero` linter would not report lints for numeric
+  values appearing in Sass script (e.g. function arguments)
 
 ## 0.10.1